teos - moves errors to common so it can be used by clients
